Rename the Features test directory to Feature
